Move DSE to new subproject
