+++ settings.gradle.kts
@@ -14,6 +14,7 @@ include(
 		"language-semantics",
 		"language-web",
 		"store",
+		"store-dse",
 		"store-query",
 		"store-query-viatra",
 		"store-reasoning",
