+++ subprojects/store-dse/src/main/java/tools/refinery/store/dse/ActionFactory.java
@@ -1,0 +1,9 @@
+package tools.refinery.store.dse;
+
+import org.eclipse.collections.api.block.procedure.Procedure;
+import tools.refinery.store.model.Model;
+import tools.refinery.store.tuple.Tuple;
+
+public interface ActionFactory {
+	Procedure<Tuple> prepare(Model model);
+}
