+++ subprojects/store-dse/src/main/java/tools/refinery/store/dse/DesignSpaceExplorationStoreAdapter.java
@@ -1,0 +1,6 @@
+package tools.refinery.store.dse;
+
+import tools.refinery.store.adapter.ModelStoreAdapter;
+
+public interface DesignSpaceExplorationStoreAdapter extends ModelStoreAdapter {
+}
