+++ subprojects/store-dse/src/main/java/tools/refinery/store/dse/Strategy.java
@@ -1,0 +1,8 @@
+package tools.refinery.store.dse;
+
+public interface Strategy {
+
+	public void initStrategy(DesignSpaceExplorationAdapter designSpaceExplorationAdapter);
+
+	public void explore();
+}
