--- subprojects/store-query-viatra/src/main/java/tools/refinery/store/query/dse/ActionFactory.java
@@ -1,9 +1,0 @@
-package tools.refinery.store.query.dse;
-
-import org.eclipse.collections.api.block.procedure.Procedure;
-import tools.refinery.store.model.Model;
-import tools.refinery.store.tuple.Tuple;
-
-public interface ActionFactory {
-	Procedure<Tuple> prepare(Model model);
-}
