--- subprojects/store-query-viatra/src/main/java/tools/refinery/store/query/dse/DesignSpaceExplorationStoreAdapter.java
@@ -1,6 +1,0 @@
-package tools.refinery.store.query.dse;
-
-import tools.refinery.store.adapter.ModelStoreAdapter;
-
-public interface DesignSpaceExplorationStoreAdapter extends ModelStoreAdapter {
-}
