--- subprojects/store-query-viatra/src/main/java/tools/refinery/store/query/dse/Strategy.java
@@ -1,10 +1,0 @@
-package tools.refinery.store.query.dse;
-
-import tools.refinery.store.model.Model;
-
-public interface Strategy {
-
-	public void initStrategy(DesignSpaceExplorationAdapter designSpaceExplorationAdapter);
-
-	public void explore();
-}
