--- subprojects/store-query-viatra/src/main/java/tools/refinery/store/query/dse/internal/Activation.java
@@ -1,9 +1,0 @@
-package tools.refinery.store.query.dse.internal;
-
-import tools.refinery.store.tuple.Tuple;
-
-public record Activation(TransformationRule transformationRule, Tuple activation) {
-	public boolean fire() {
-		return transformationRule.fireActivation(activation);
-	}
-}
